+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.3.0 / 2018-06-08
 - Expose models name as constants
+- Add get models helper
 
 # v0.1.0 / 2018-05-22